feat: vls checker extends Checker
