+++ packages/checker-vls/src/main.ts
@@ -1,84 +1,9 @@
-import {
-  CreateDiagnostic,
-  createScript,
-  ServeAndBuildChecker,
-  SharedConfig,
-} from 'vite-plugin-checker'
-import { isMainThread, parentPort } from 'worker_threads'
+import * as commander from 'commander'
 
-import { DiagnosticOptions, diagnostics } from './commands/diagnostics'
-
-import type { ConfigEnv } from 'vite'
-
-export const createDiagnostic: CreateDiagnostic = (userOptions = {}) => {
-  let overlay = true // Vite defaults to true
-
-  return {
-    config: ({ hmr }) => {
-      const viteOverlay = !(typeof hmr === 'object' && hmr.overlay === false)
-
-      if (userOptions.overlay === false || !viteOverlay) {
-        overlay = false
-      }
-    },
-    async configureServer({ root }) {
-      const workDir: string = userOptions.root ?? root
-      const errorCallback: DiagnosticOptions['errorCallback'] = (diagnostics, overlayErr) => {
-        if (!overlay) return
-        if (!overlayErr) return
-
-        parentPort?.postMessage({
-          type: 'ERROR',
-          payload: {
-            type: 'error',
-            err: overlayErr,
-          },
-        })
-      }
-
-      await diagnostics(workDir, 'WARN', { watch: true, errorCallback, verbose: false })
-    },
-  }
-}
-
-const { mainScript, workerScript } = createScript<{ vls: VlsConfig }>({
-  absFilename: __filename,
-  buildBin: ['vite-plugin-checker-vls', ['diagnostics']],
-  serverChecker: { createDiagnostic },
-})!
-
-if (isMainThread) {
-  const configCurryFn = (vlsConfig: VlsConfig) => {
-    return (sharedConfig: SharedConfig, env: ConfigEnv) => {
-      const createChecker = mainScript()
-      return createChecker({ vls: vlsConfig, ...sharedConfig }, env)
-    }
-  }
-
-  module.exports.VlsChecker = configCurryFn
-  module.exports.createServeAndBuild = configCurryFn
-} else {
-  workerScript()
-}
-
-type VlsConfig = Partial<{
-  // TODO: support VLS config
-}>
-
-declare const VlsChecker: (
-  options?: VlsConfig
-) => (config: VlsConfig & SharedConfig) => ServeAndBuildChecker
-
-export { VlsChecker }
-export type { VlsConfig }
-
-import commander from 'commander'
-
-export * as vscodeLanguageserverProtocol from 'vscode-languageserver-protocol'
-export * as vscodeUri from 'vscode-uri'
+export { commander }
 export * as chokidar from 'chokidar'
 export * as vls from 'vls'
-export { commander }
-
 export * as vscodeLanguageserverNode from 'vscode-languageserver/node'
+export * as vscodeLanguageserverProtocol from 'vscode-languageserver-protocol'
 export * as vscodeLanguageserverTextdocument from 'vscode-languageserver-textdocument'
+export * as vscodeUri from 'vscode-uri'
