+++ playground/vue2-vls/vite.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vite'
 import { createVuePlugin } from 'vite-plugin-vue2'
 import ViteComponents from 'vite-plugin-components'
 import Checker from 'vite-plugin-checker'
-import { VlsChecker } from 'vite-plugin-checker-vls'
+// import { VlsChecker } from 'vite-plugin-checker-vls'
 import { resolve } from 'path'
 
 const config = defineConfig({
@@ -18,7 +18,8 @@ const config = defineConfig({
   plugins: [
     createVuePlugin({}),
     ViteComponents({ transformer: 'vue2' }),
-    Checker({ vls: VlsChecker() }),
+    // @ts-ignore
+    Checker({ vls: {} }),
   ],
   server: {
     port: 8080,
